+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
 
 * Consolidated Named & FluentName update mod.rs.
-* Adding encoding_rs to restrict named characters to UTF-8.
+* Adding validation of the FluentName String to alphanumeric or hyphens.
+* Updating dependency version numbers to only use MAJOR and MINOR and not PATCH to avoid having to constantly update.
+* Adding log warning to when FluentName defaults to blank bacause of bad name.
+* I love the `thiserror` crate. 
+